Update 4
Probably last and final update. Kinda changed the layout. Now you can start a service that works in the background. It will save your lat and long and the show them to you when you open the app. There is a "Clear data" button too, which deletes the data from your screen AND from your device.